pyproject.toml
[project]
name = "notestonotion"
version = "0.1.0"
description = "ETL pipeline for scientific handwritten notes - PDF to Notion with LaTeX support"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "NotesToNotion" }
]
dependencies = [
    "google-generativeai>=0.8.0",
    "notion-client>=2.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
notestonotion = "main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
